requirements.txt
backports-abc==0.5
backports.functools-lru-cache==1.5
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cloudpickle==1.1.1
cycler==0.10.0
dask==1.2.2
decorator==4.4.0
easydict==1.9
Flask==1.1.1
flask-redis==0.4.0
Flask-SocketIO==4.2.1
funcsigs==1.0.2
future==0.18.2
futures==3.2.0
graphviz==0.8.4
idna==2.8
imutils==0.5.2
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==2.2.4
mock==3.0.5
mxnet-cu100mkl==1.4.0.post0
networkx==2.2
numexpr==2.6.9
numpy==1.14.6
onnx==1.2.1
opencv-contrib-python==4.1.0.25
opencv-python==4.1.2.30
pandas==0.24.2
Pillow==6.0.0
protobuf==3.7.1
pybase64==0.5.0
pyparsing==2.4.0
python-dateutil==2.8.0
python-engineio==3.11.1
python-socketio==4.4.0
pytz==2019.1
PyWavelets==1.0.3
redis==3.3.11
requests==2.21.0
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
singledispatch==3.4.0.3
six==1.12.0
subprocess32==3.5.4
tables==3.5.1
toolz==0.9.0
tornado==5.1.1
typing==3.6.6
typing-extensions==3.7.2
urllib3==1.24.3
Werkzeug==0.16.0
